Added alerts before changes and deletions

+++ src/components/contacts/Contact.js
@@ -1,4 +1,5 @@
 import React from "react";
+import swal from "sweetalert";
 
 import { useToggle } from "../../reducers/reusable";
 
@@ -15,11 +16,22 @@ const Contact = ({ contact, CONTACTS_URL }) => {
 
   const handleDeleteContact = (event) => {
     event.preventDefault();
-    fetch(CONTACTS_URL, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json", Authorization: localToken, userId: localId, contactId: contactId },
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this contact.",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
     })
-      .then(window.location.reload())
+      .then((willDelete) => {
+        if (willDelete) {
+          fetch(CONTACTS_URL, {
+            method: "DELETE",
+            headers: { "Content-Type": "application/json", Authorization: localToken, userId: localId, contactId: contactId },
+          })
+            .then(window.location.reload())
+        }
+      })
   };
 
   return (
@@ -38,8 +50,8 @@ const Contact = ({ contact, CONTACTS_URL }) => {
             <StyledCardText>{contact.contactPhoneNumber}</StyledCardText>
             <StyledCardText left>Relation:</StyledCardText>
             <StyledCardText>{contact.contactCategory}</StyledCardText>
-            <StyledCardButton onClick={toggleEditMode}>Edit</StyledCardButton>
             <StyledCardButton onClick={handleDeleteContact}>Delete</StyledCardButton>
+            <StyledCardButton onClick={toggleEditMode}>Edit</StyledCardButton>
           </StyledCardWithGrid>
         )
       }
